# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FlightMovement API Documentation",
        "description": "API documentation for FlightMovement",
        "contact": {
            "email": "dev@camranh.aero"
        },
        "version": "1.0.0"
    },
    "servers": [
        {

            "url": "http://localhost:8000/api",
            "description": "Rest API"
        }
    ],
    "paths": {
...
        "/aircrafts": {
            "get": {
                "tags": [
                    "Aircrafts"
                ],
                "summary": "Get all active aircrafts",
                "description": "Retrieve all active aircrafts ordered by registration",
                "operationId": "e60db6abb3ab423ff473d4b32cbd8877",
                "responses": {
                    "200": {
                        "description": "Aircrafts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "AircraftId": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "Registration": {
                                                        "type": "string",
                                                        "example": "VN-A321"
                                                    },
                                                    "AircraftTypeId": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "AirlineId": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "Status": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Aircrafts"
                ],
                "summary": "Create new aircraft",
                "description": "Create a new aircraft record",
                "operationId": "d2f24dd492a47e08149f83a3b176c638",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "AircraftId",
                                    "Registration",
                                    "AircraftTypeId",
                                    "AirlineId"
                                ],
                                "properties": {
                                    "AircraftId": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "Registration": {
                                        "type": "string",
                                        "example": "VN-A350"
                                    },
                                    "AircraftTypeId": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "AirlineId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "Status": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    },
                                    "MaxPax": {
                                        "type": "integer",
                                        "example": 300
                                    },
                                    "Mtow": {
                                        "type": "integer",
                                        "example": 280000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Aircraft created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aircraft created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/aircrafts/registration/{registration}": {
            "get": {
                "tags": [
                    "Aircrafts"
                ],
                "summary": "Get aircraft by registration",
                "description": "Retrieve aircraft information by registration number",
                "operationId": "caa65704a4e927371368f6f4142b28e3",
                "parameters": [
                    {
                        "name": "registration",
                        "in": "path",
                        "description": "Aircraft registration (e.g., VN-A321)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "VN-A321"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aircraft found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "AircraftId": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "Registration": {
                                                    "type": "string",
                                                    "example": "VN-A321"
                                                },
                                                "AircraftTypeId": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "AirlineId": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aircraft not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aircraft not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/aircrafts/{aircraftId}": {
            "put": {
                "tags": [
                    "Aircrafts"
                ],
                "summary": "Update aircraft",
                "description": "Update an existing aircraft record",
                "operationId": "518efb71ee9c3f527ba1d3e571f845f9",
                "parameters": [
                    {
                        "name": "aircraftId",
                        "in": "path",
                        "description": "Aircraft ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Registration": {
                                        "type": "string",
                                        "example": "VN-A321-Updated"
                                    },
                                    "Status": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    },
                                    "MaxPax": {
                                        "type": "integer",
                                        "example": 180
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aircraft updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aircraft updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aircraft not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aircraft not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}